Query tells state manager which relationships have been fixed up
Issue #5688

Query code will take care of fixing up entities within the query when Include is used. This means that the state manager does not need to do fixup for these relationships in cases where there are no entities tracked other than those that have come from query. This change allows query to communicate this to the state manager so that it can avoid unneeded work. The change also preserves the very simple case of a single query for a single entity type with no self-refs, in which case the state manager does not need to do work at all.